# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lineserve API",
    "description": "API for managing OpenStack resources",
    "version": "1.0.0",
    "contact": {
      "name": "Lineserve Support",
      "url": "https://lineserve.org"
    }
  },
  "servers": [
    {
      "url": "https://api.lineserve.org/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Image ID"
          },
          "name": {
            "type": "string",
            "description": "Image name"
          },
          "status": {
            "type": "string",
            "description": "Image status"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Image size in bytes"
          },
          "visibility": {
            "type": "string",
            "description": "Image visibility (private, public, shared, community)"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Image tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Image properties"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
...
    "/images": {
      "get": {
        "summary": "List all images",
        "tags": ["images"],
        "description": "Returns a list of all available images",
        "operationId": "listImages",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Image"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload a new image",
        "tags": ["images"],
        "description": "Creates a new image and optionally uploads image data",
        "operationId": "createImage",
        "requestBody": {
          "description": "Image creation request",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file to upload"
                  },
                  "name": {
                    "type": "string",
                    "description": "Name of the image"
                  },
                  "container_format": {
                    "type": "string",
                    "description": "Container format (e.g., bare, ovf)",
                    "default": "bare"
                  },
                  "disk_format": {
                    "type": "string",
                    "description": "Disk format (e.g., raw, qcow2, vhd)",
                    "default": "raw"
                  },
                  "min_disk": {
                    "type": "integer",
                    "description": "Minimum disk size in GB"
                  },
                  "min_ram": {
                    "type": "integer",
                    "description": "Minimum RAM in MB"
                  },
                  "protected": {
                    "type": "boolean",
                    "description": "Whether the image is protected"
                  },
                  "visibility": {
                    "type": "string",
                    "description": "Visibility of the image (private, public, shared, community)",
                    "default": "private"
                  }
                },
                "required": ["name"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Image created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/images/{id}": {
      "get": {
        "summary": "Get image details",
        "tags": ["images"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Image details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "404": {
            "description": "Image not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "images"
        ],
        "summary": "Delete an image",
        "description": "Deletes an image by ID",
        "operationId": "deleteImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Image not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
} 